[web] Add RowActions component
