+++ web/src/components/users/FirstUser.jsx
@@ -27,39 +27,15 @@ import {
   Alert,
   Button,
   Checkbox,
-  DropdownToggle,
   Form,
   FormGroup,
   TextInput,
   Skeleton,
 } from "@patternfly/react-core";
 
-import { TableComposable, Thead, Tr, Th, Tbody, Td, ActionsColumn } from '@patternfly/react-table';
+import { TableComposable, Thead, Tr, Th, Tbody, Td } from '@patternfly/react-table';
 
-import { Icon } from '~/components/layout';
-import { PasswordAndConfirmationInput, Popup } from '~/components/core';
-
-const RowActions = ({ actions, id, ...props }) => {
-  const actionsToggle = (props) => (
-    <DropdownToggle
-      id={id}
-      aria-label="Actions"
-      toggleIndicator={null}
-      isDisabled={props.isDisabled}
-      onToggle={props.onToggle}
-    >
-      <Icon name="more_vert" size="24" />
-    </DropdownToggle>
-  );
-
-  return (
-    <ActionsColumn
-      items={actions}
-      actionsToggle={actionsToggle}
-      {...props}
-    />
-  );
-};
+import { RowActions, PasswordAndConfirmationInput, Popup } from '~/components/core';
 
 const UserNotDefined = ({ actionCb }) => {
   return (
